+++ persistence/migrations/0090_pollAnswers.ts
@@ -12,7 +12,7 @@ export const up = async (knex: Knex): Promise<void> =>
     table.integer("score");
 
     table.boolean("deleted").nullable();
-    table.integer("createdAt");
+    table.bigInteger("createdAt");
     table.integer("updatedAt").nullable();
 
     table.foreign("pollQuestionId").references("id").inTable("pollQuestions");
