+++ persistence/migrations/0120_userSessions.ts
@@ -16,7 +16,7 @@ export const up = async (knex: Knex): Promise<void> =>
     );
 
     table.boolean("deleted");
-    table.integer("createdAt");
+    table.bigInteger("createdAt");
     table.integer("updatedAt");
 
     table.foreign("userId").references("id").inTable("users");
