+++ persistence/migrations/0130_invites.ts
@@ -10,7 +10,7 @@ export const up = async (knex: Knex): Promise<void> =>
     table.string("phone").nullable();
 
     table.boolean("deleted").nullable();
-    table.integer("createdAt");
+    table.bigInteger("createdAt");
     table.integer("updatedAt").nullable();
 
     table.foreign("sampleGroupId").references("id").inTable("sampleGroups");
